community_bot/requirements.txt
python-telegram-bot==21.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
Pillow>=10.0.0
